requirements.txt
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
comm==0.1.2
cryptography==40.0.2
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.9.0
Flask==3.0.2
fqdn==1.5.1
frozendict==2.3.7
html5lib==1.1
idna==3.4
imageio==2.34.0
instaloader==4.9.6
ipython==8.11.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
lazy_loader==0.3
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.5
multitasking==0.0.11
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.2.1
numpy==1.24.3
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.1.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pyarrow==15.0.0
pycparser==2.21
Pygments==2.14.0
pylance==0.10.4
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.1
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.3.0
tifffile==2024.2.12
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.14
virtualenv==20.20.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==3.0.1
yfinance==0.2.18
